Creates token secret for service account
This patch changes the default behavior of waiting for the generation of
a default secret for the service account and propagate that over to the
guest cluster.
Instead, alongwith the service account a secret of type
"kubernetes.io/service-account-token" is created to host the token for
the service account. The controller waits for the realization of this
secret and then propagates this data over to the guest cluster.

This is required since 1.24 disables the auto generation of a default
token secret for every service account created.

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>
(cherry picked from commit ec40a83e830db39323e4bbba73ddb27e443bc3e4)